Move common k8s config to a separate package
