+++ component/common/kubernetes/kubernetes_test.go
@@ -1,4 +1,4 @@
-package config
+package kubernetes
 
 import (
 	"testing"
