+++ component/loki/source/kubernetes_events/kubernetes_events.go
@@ -15,6 +15,7 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/common/config"
+	"github.com/grafana/agent/component/common/kubernetes"
 	"github.com/grafana/agent/component/common/loki"
 	"github.com/grafana/agent/component/common/loki/positions"
 	"github.com/grafana/agent/pkg/river"
@@ -46,7 +47,7 @@ type Arguments struct {
 	Namespaces []string `river:"namespaces,attr,optional"`
 
 	// Client settings to connect to Kubernetes.
-	Client config.ClientArguments `river:"client,block,optional"`
+	Client kubernetes.ClientArguments `river:"client,block,optional"`
 }
 
 var _ river.Unmarshaler = (*Arguments)(nil)
@@ -55,7 +56,7 @@ var _ river.Unmarshaler = (*Arguments)(nil)
 var DefaultArguments = Arguments{
 	JobName: "loki.source.kubernetes_events",
 
-	Client: config.ClientArguments{
+	Client: kubernetes.ClientArguments{
 		HTTPClientConfig: config.DefaultHTTPClientConfig,
 	},
 }
